+++ ui/shared/components/panel/search/FrequencyFilter.jsx
@@ -30,6 +30,14 @@ const AF_STEP_LABELS = {
 
 const AF_OPTIONS = AF_STEPS.map(value => ({ value }))
 
+const updateFrequency = ({ onChange, field, initialValue, nullField }) => (val) => {
+  const updateFields = { [field]: val }
+  if (nullField) {
+    updateFields[nullField] = null
+  }
+  onChange({ ...initialValue, ...updateFields })
+}
+
 const FrequencyIntegerInput = React.memo(({ label, value, field, nullField, inlineAF, onChange }) => (
   <IntegerInput
     label={label}
@@ -37,13 +45,7 @@ const FrequencyIntegerInput = React.memo(({ label, value, field, nullField, inli
     min={0}
     max={100}
     width={inlineAF ? 3 : 8}
-    onChange={(val) => {
-      const updateFields = { [field]: val }
-      if (nullField) {
-        updateFields[nullField] = null
-      }
-      onChange({ ...value, ...updateFields })
-    }}
+    onChange={updateFrequency({ onChange, field, initialValue: value, nullField })}
   />
 ))
 
@@ -59,7 +61,7 @@ FrequencyIntegerInput.propTypes = {
 export const AfFilter = ({ value, onChange, inline, label, width }) => {
   const afProps = {
     value: (value || {}).af,
-    onChange: val => onChange({ ...value, af: val, ac: null }),
+    onChange: updateFrequency({ onChange, initialValue: value, field: 'af', nullField: 'ac' }),
   }
   return inline ?
     <Select options={AF_OPTIONS} width={width || 5} label={label || 'AF'} {...afProps} /> :
@@ -113,15 +115,21 @@ const formatHeaderValue = values => Object.values(values).reduce((acc, value) =>
   hh: value.hh === acc.hh ? value.hh : null,
 }), Object.values(values)[0])
 
+const callsetChange = (onChange, initialValues) => val => onChange(
+  { ...initialValues, [THIS_CALLSET_FREQUENCY]: val, [SV_CALLSET_FREQUENCY]: val },
+)
+
+const freqChange = (onChange, initialValues) => val => onChange(FREQUENCIES.filter(
+  ({ name }) => name !== THIS_CALLSET_FREQUENCY && name !== SV_CALLSET_FREQUENCY,
+).reduce((acc, { name }) => ({ ...acc, [name]: val }), initialValues || {}))
+
 export const HeaderFrequencyFilter = ({ value, onChange, ...props }) => {
   const { callset, sv_callset: svCallset, ...freqValues } = value || {}
   const headerValue = freqValues ? formatHeaderValue(freqValues) : {}
 
-  const onCallsetChange = val => onChange({ ...freqValues, [THIS_CALLSET_FREQUENCY]: val, [SV_CALLSET_FREQUENCY]: val })
+  const onCallsetChange = callsetChange(onChange, freqValues)
 
-  const onFreqChange = val => onChange(FREQUENCIES.filter(
-    ({ name }) => name !== THIS_CALLSET_FREQUENCY && name !== SV_CALLSET_FREQUENCY,
-  ).reduce((acc, { name }) => ({ ...acc, [name]: val }), value || {}))
+  const onFreqChange = freqChange(onChange, value)
 
   return (
     <FrequencyFilter {...props} value={headerValue} onChange={onFreqChange} homHemi inlineAF>