+++ ui/shared/components/panel/variants/FamilyVariantTags.jsx
@@ -28,6 +28,10 @@ const TagTitle = styled.span`
   color: #999;
 `
 
+const RedItal = styled.i`
+  color: red;
+`
+
 const NO_DISPLAY = { display: 'none' }
 
 const SHORTCUT_TAGS = ['Review', 'Excluded']
@@ -37,7 +41,7 @@ const VARIANT_NOTE_FIELDS = [{
   label: (
     <label>
       Add to
-      <i style={{ color: 'red' }}>&nbsp; ClinVar &nbsp;</i>
+      <RedItal>&nbsp; ClinVar &nbsp;</RedItal>
       submission
     </label>
   ),
@@ -77,6 +81,8 @@ export const taggedByPopup = (tag, title) => (trigger, hideMetadata) => (
     }
   />
 )
+
+const notePopup = note => note && taggedByPopup(note, 'Note By')
 
 const ShortcutTagToggle = React.memo(({ tag, ...props }) => {
   const toggle = <InlineToggle color={tag && tag.color} divided {...props} value={tag} />
@@ -256,7 +262,7 @@ const FamilyVariantTags = React.memo(({
               isEditable
               showInLine
               compact
-              getTextPopup={note => note && taggedByPopup(note, 'Note By')}
+              getTextPopup={notePopup}
               onSubmit={dispatchUpdateVariantNote}
             />
           </Table.Cell>